moved SetSizeFrame class into own file
